# Conflicts:
#	app/declarations/react-native-button.d.ts
#	app/declarations/zxcvbn.d.ts

+++ app/components/Views/Settings/Contacts/index.tsx
@@ -1,10 +1,15 @@
-import React, { PureComponent } from 'react';
+import React, { ComponentType, PureComponent } from 'react';
 import { SafeAreaView, StyleSheet } from 'react-native';
-import PropTypes from 'prop-types';
 import { strings } from '../../../../../locales/i18n';
 import { getNavigationOptionsTitle } from '../../../UI/Navbar';
 import { connect } from 'react-redux';
+import { NavigationProp, ParamListBase } from '@react-navigation/native';
+import { Hex } from '@metamask/utils';
+import { AddressBookControllerState } from '@metamask/address-book-controller';
+import { RootState } from '../../../../reducers';
+import { Colors, Theme } from '../../../../util/theme/models';
 import AddressList from '../../confirmations/legacy/SendFlow/AddressList';
+import { AddressListProps } from '../../confirmations/legacy/SendFlow/AddressList/AddressList.types';
 import StyledButton from '../../../UI/StyledButton';
 import Engine from '../../../../core/Engine';
 import ActionSheet from '@metamask/react-native-actionsheet';
@@ -15,7 +20,7 @@ import Routes from '../../../../../app/constants/navigation/Routes';
 import { ContactsViewSelectorIDs } from '../../../../../e2e/selectors/Settings/Contacts/ContacsView.selectors';
 import { selectAddressBook } from '../../../../selectors/addressBookController';
 
-const createStyles = (colors) =>
+const createStyles = (colors: Colors) =>
   StyleSheet.create({
     wrapper: {
       backgroundColor: colors.background.default,
@@ -34,32 +39,47 @@ const ADD = 'add';
 /**
  * View that contains app information
  */
-class Contacts extends PureComponent {
-  static propTypes = {
-    /**
-     * Map representing the address book
-     */
-    addressBook: PropTypes.object,
-    /**
-     /* navigation object required to push new views
-     */
-    navigation: PropTypes.object,
-    /**
-     * The chain ID for the current selected network
-     */
-    chainId: PropTypes.string,
-  };
+const TypedAddressList = AddressList as ComponentType<
+  AddressListProps & { chainId: Hex }
+>;
 
-  state = {
+interface ContactsProps {
+  /**
+   * Map representing the address book
+   */
+  addressBook: AddressBookControllerState['addressBook'];
+  /**
+   * navigation object required to push new views
+   */
+  navigation: NavigationProp<ParamListBase>;
+  /**
+   * The chain ID for the current selected network
+   */
+  chainId: Hex;
+}
+
+interface ContactsState {
+  reloadAddressList: boolean;
+}
+
+interface ActionSheetRef {
+  show: () => void;
+}
+
+class Contacts extends PureComponent<ContactsProps, ContactsState> {
+  static contextType = ThemeContext;
+
+  state: ContactsState = {
     reloadAddressList: false,
   };
 
-  actionSheet;
-  contactAddressToRemove;
+  actionSheet: ActionSheetRef | null = null;
+  contactAddressToRemove: string | undefined;
 
   updateNavBar = () => {
     const { navigation } = this.props;
-    const colors = this.context.colors || mockTheme.colors;
+    const colors =
+      (this.context as Theme | undefined)?.colors || mockTheme.colors;
     navigation.setOptions(
       getNavigationOptionsTitle(
         strings('app_settings.contacts_title'),
@@ -74,7 +94,7 @@ class Contacts extends PureComponent {
     this.updateNavBar();
   };
 
-  componentDidUpdate = (prevProps) => {
+  componentDidUpdate = (prevProps: ContactsProps) => {
     this.updateNavBar();
     const { chainId } = this.props;
     if (
@@ -93,7 +113,7 @@ class Contacts extends PureComponent {
     }, 100);
   };
 
-  onAddressLongPress = (address) => {
+  onAddressLongPress = (address: string) => {
     this.contactAddressToRemove = address;
     this.actionSheet && this.actionSheet.show();
   };
@@ -101,11 +121,13 @@ class Contacts extends PureComponent {
   deleteContact = () => {
     const { AddressBookController } = Engine.context;
     const { chainId } = this.props;
-    AddressBookController.delete(chainId, this.contactAddressToRemove);
+    if (this.contactAddressToRemove !== undefined) {
+      AddressBookController.delete(chainId, this.contactAddressToRemove);
+    }
     this.updateAddressList();
   };
 
-  onAddressPress = (address) => {
+  onAddressPress = (address: string) => {
     this.props.navigation.navigate('ContactForm', {
       mode: EDIT,
       editMode: EDIT,
@@ -118,7 +140,7 @@ class Contacts extends PureComponent {
     this.props.navigation.navigate('ContactForm', { mode: ADD });
   };
 
-  createActionSheetRef = (ref) => {
+  createActionSheetRef = (ref: ActionSheetRef | null) => {
     this.actionSheet = ref;
   };
 
@@ -131,8 +153,9 @@ class Contacts extends PureComponent {
 
   render = () => {
     const { reloadAddressList } = this.state;
-    const colors = this.context.colors || mockTheme.colors;
-    const themeAppearance = this.context.themeAppearance;
+    const theme = this.context as Theme | undefined;
+    const colors = theme?.colors || mockTheme.colors;
+    const themeAppearance = theme?.themeAppearance;
     const styles = createStyles(colors);
     const { chainId } = this.props;
 
@@ -141,7 +164,7 @@ class Contacts extends PureComponent {
         style={styles.wrapper}
         testID={ContactsViewSelectorIDs.CONTAINER}
       >
-        <AddressList
+        <TypedAddressList
           chainId={chainId}
           onlyRenderAddressBook
           reloadAddressList={reloadAddressList}
@@ -167,7 +190,9 @@ class Contacts extends PureComponent {
           cancelButtonIndex={1}
           destructiveButtonIndex={0}
           // eslint-disable-next-line react/jsx-no-bind
-          onPress={(index) => (index === 0 ? this.deleteContact() : null)}
+          onPress={(index: number) =>
+            index === 0 ? this.deleteContact() : null
+          }
           theme={themeAppearance}
         />
       </SafeAreaView>
@@ -175,11 +200,11 @@ class Contacts extends PureComponent {
   };
 }
 
-Contacts.contextType = ThemeContext;
-
-const mapStateToProps = (state) => ({
+const mapStateToProps = (state: RootState) => ({
   addressBook: selectAddressBook(state),
-  chainId: selectChainId(state),
+  // AddressBookController is EVM-only; on non-EVM networks the selected
+  // chain id is a CAIP id and address book lookups yield no entries.
+  chainId: selectChainId(state) as Hex,
 });
 
 export default connect(mapStateToProps)(Contacts);
